Add status check for the current user upon each action & rerender

+++ src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import { BsFillLockFill, BsFillUnlockFill, BsFillTrashFill, BsDoorOpenFill } from "react-icons/bs"
 import { useState, useEffect } from "react"
 import { nanoid } from "nanoid"
-import { useNavigate } from "react-router"
+import { useNavigate, Navigate } from "react-router"
 import { signOut } from "firebase/auth"
 import { auth } from "../firebase"
 // db management
@@ -46,26 +46,61 @@ export default function Home() {
         console.log("Table Population effect fired")
     }, [])
 
-    async function deleteSelection() {
-        if (selectedRows && selectedRows.length > 0) {
-            for (let i = 0; i < selectedRows.length; i++) {
-                try {
-                    await deleteDoc(doc(db, "users", selectedRows[i]))
-                    console.log(`Doc deleted`)
-                    console.log(selectedRows)
-                } catch (error) {
-                    console.error(`Error with deletion: ${error}`)
-                }
+
+    async function checkIfAccountActive() {
+        try {
+            const snapshot = await getDocs(usersCollectionRef)
+            const users = snapshot.docs.map(doc => ({ ...doc.data(), id: doc.id }))
+            if (!auth.currentUser) {
+                return false
+            } else if (!users.filter(user => user.id === auth.currentUser.uid)[0]) {
+                return false
+            } else if (users.filter(user => user.id === auth.currentUser.uid)[0].status === "blocked") {
+                return false
+            } else {
+                return true
             }
-            setSelectedRows([])
-            populateTable()
+        } catch (error) {
+            console.error(`Error with getting a snapshot in checkAccountStatus: ${error} `)
         }
-        else {
-            console.log("Nothing has been selected")
+    }
+
+    useEffect(() => {
+        async function checkIfAccountActiveOnPopulate() {
+            const accountActive = await checkIfAccountActive()
+            console.log(`accountActive = ${accountActive}`)
+            if (!accountActive) {
+                await signOut(auth)
+            }
+        }
+        checkIfAccountActiveOnPopulate()
+    }, [tableData])
+
+    async function deleteSelection() {
+        const accountActive = await checkIfAccountActive()
+        if (accountActive) {
+            if (selectedRows && selectedRows.length > 0) {
+                for (let i = 0; i < selectedRows.length; i++) {
+                    try {
+                        await deleteDoc(doc(db, "users", selectedRows[i]))
+                        console.log(`Doc deleted`)
+                    } catch (error) {
+                        console.error(`Error with deletion: ${error}`)
+                    }
+                }
+                setSelectedRows([])
+                populateTable()
+            } else {
+                console.log("Nothing has been selected")
+            }
+        } else if (!accountActive) {
+            await signOut(auth)
+            navigate("/authorize")
         }
     }
 
     async function blockSelection() {
+        checkIfAccountActive()
         if (selectedRows && selectedRows.length > 0) {
             for (let i = 0; i < selectedRows.length; i++) {
                 try {
@@ -83,6 +118,7 @@ export default function Home() {
     }
 
     async function unblockSelection() {
+        checkIfAccountActive()
         if (selectedRows && selectedRows.length > 0) {
             for (let i = 0; i < selectedRows.length; i++) {
                 try {
@@ -103,6 +139,12 @@ export default function Home() {
     if (!tableData) {
         return (
             <p>Loading...</p>
+        )
+    }
+
+    if (!auth.currentUser) {
+        return (
+            <Navigate to="/authorize"></Navigate>
         )
     }
 
@@ -175,7 +217,7 @@ export default function Home() {
                                         className="Home__Checkbox"
                                     />
                                 </td>
-                                <td>{entry.name}</td>
+                                <td>{entry.name} {entry.id === auth.currentUser.uid ? "(you)" : ""}</td>
                                 <td>{entry.email}</td>
                                 <td>{entry.lastSeen}</td>
                                 <td>{entry.status[0].toUpperCase() + entry.status.slice(1)}</td>
